data/covid-19-data-master 16.5.22/scripts/requirements.txt
beautifulsoup4~=4.9.0
boto3==1.18.43
click~=8.0.0
dateparser~=1.1.0
epiweeks~=2.1.0
gdown~=4.2.0
gsheets==0.5.1
joblib==1.0.1
lxml>=4.6.3
numpy~=1.21.0
odfpy~=1.4.1
openpyxl==3.0.7
pandas~=1.3.0
pdfminer.six==20211012
Pillow==9.0.1
psutil~=5.9.0
py-cpuinfo~=8.0.0
pyaml-env~=1.1.0
PyDrive~=1.3.0
PyMySQL==0.9.3
pyperclip==1.8.2
python-dateutil~=2.8.0
python-dotenv~=0.18.0
pytz==2021.3
PyYAML==5.4.1
requests~=2.27.0
selenium==3.141.0
slackclient==2.9.3
TableauScraper~=0.1.0
tabula-py==2.2.0
tabulate==0.8.9
termcolor==1.1.0
tqdm==4.61.1
tweepy~=3.10.0
uk-covid19~=1.2.0
Unidecode~=1.2.0
xlrd==2.0.1
xlsx2csv==0.7.8
XlsxWriter==1.4.3
